Fixed bug and Write out file
I found a bug in the File_read.cpp(in some cases it was not reading in all the ingredients). I fixed that and I am adding the write out function for the recipe file. in this function is the key to deleting a recipe. to delete a recipe change the recipe name to DELETE and it will not be printed to file. I have not tested it yet but it should work.